Move litho libs to /lib
