+++ litho/scripts/Lithographer.py
@@ -3,10 +3,15 @@ from tkinter.ttk import Progressbar
 from collections import namedtuple
 from PIL import  Image
 from time import sleep
-from litho_img_lib import *
-from litho_gui_lib import *
 from stage_control.grbl_stage import GrblStage
 import serial
+
+#import sys and use path insert to add lib files
+import sys
+from os.path import join, dirname, realpath
+sys.path.insert(0, join(dirname(dirname(realpath(__file__))), "lib"))
+from litho_gui_lib import *
+from litho_img_lib import *
 
 from config import camera as camera_hw
 import cv2
